Bugfixes and cleanup
- Fix an integer sign bug causing custom SFX not to play
- Remove ugly using directives
- Make the sfx directory a configurable setting
- Fix some comments

+++ game/sfxreplace/Dvd.cpp
@@ -25,8 +25,8 @@ bool DVDOpenDir(const char* dirName, DVDDir* dir) {
     if (!entryIsDir(entry))
         return false;
 
-    dir->entryNum = (u32)entry;
-    dir->location = (u32)entry + 1;
+    dir->entryNum = entry;
+    dir->location = entry + 1;
     dir->next = nextDir(entry);
     return true;
 }
